Broke a test in IDesignSpaceTest.java
testNegativeNextParameterConstructorOverride:
An InPUTException seems to be semi-expected (it is caught, but the absense of
such an exception doesn't lead to failure). The test then goes on to check
the type of the returned value. An InPUTException is indeed thrown, which
may or may not be intended. If it is intended, then the type test would seem
to be irrelevant.

The test now throws the InPUTException and fails.