Cleaned and extended a test in IDesignTest.java
Added a helper method setAndCompare that takes a value object and a parameter
id. It tests that setting and then getting the parameter yields the original
value, with the additional constraint that the test value must be different
from any value that is already set.
testSetPrimitive uses the new helper method, making it cleaner.