Fixed a test and a possible bug in Range
The IDesignTest.testSetRelativePrimitives checks that an out of range value
will not be accepted. The catch(Exception) was hiding a NullPointerException
that was being thrown while performing range checking.
Trying to fix that error seemed to reveal a bug in the range checking itself.
Based on the test case and the exception that is generated in response to the
range violation, it looked like the comparison was reversed.
The test has been changed to expect the IllegalArgumentException, but not an
InPUTException.
These changes make the test pass.